Bring up to 2.7.0rc2
mfc subtheme was outdated -- i've long since moved from a text/bg approach to a foreground/background approach. i need to modularize this a bit more though... that should make it easier to copy from the css to the scss, and also to allow admins to choose which mods to selectively apply, same as users can do on the userstyles.org version of mfc.